requirements.txt
anthropic>=0.84.0
google-genai>=1.66.0
python-gitlab>=8.1.0
PyGithub>=2.8.0
pytest>=9.0.2
pytest-mock>=3.15.0
protobuf>=5.29.0 #Latest 7.34.0
